Only Process Unique URIs for Cleaning
Fix a bug where a sources with the same URI were each being processed during the Clean process, causing excess recursion down the same tree, and leading to files / directories being added to the list for cleaning multiple times, leading to an error when attempting to be removed a subsequent time.

Changed to get a unique list of URIs and process those only.