Process only Modified Indexes
Code will now timestamp before an after of all Release files it intends to download in order to detect whether they have changed. If they have changed, then the files will be processed as normal. If they are the same, (or not downloaded as they did not exist in the mirror), then the files are left untouched. This leads to greatly reduced processing times for regular runs of the application.

Additionally, files copied from Skel to Mirror are only done so if the Skel files have a newer timestamp than that of the mirror file (if it exists)