Further Fixes and Improvements for Source Processing Optimisation
Fix bug where Compression Format not being added for Contents files.
Change Getting of Release Files to get Modified or Unmodified files.
Adjust Regex for __ProcessLine to not start from start of line.
Only build file list for Modified sources.
Remove duplicate files from downloads.
Adjust copying of Skel to Mirror to only copy newer files.
Overhaul Clean function so that unmodified sources for the URI are processed and added to filesToKeep to ensure they are not removed if not processed. Change to using os.walk(). Not sure if faster, but adds some form of output for the user.