Address #29 about missing files
Due to an older issue, where refrapt was being interrupted, only parts a repository were downloaded, while the Index file was the newest. The change detection therefore prevented the missing files form being downloaded. A setting already exists to ignore the change detection, and force download anyway, but it was not applied everywhere it needed to be. It is now applied to the change detection for Index files, so all files will be added regardless